% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_e__enginebuilder_LLM_train__mwiller.R
\name{build_llm_template_train}
\alias{build_llm_template_train}
\title{LLM Prompt Template for Training Engine}
\usage{
build_llm_template_train(task_description)
}
\arguments{
\item{task_description}{Character. The user-facing description of the training logic.}
}
\value{
A character string representing the LLM-ready prompt.
}
\description{
Generates an engine-specific prompt for training engines.
}
