% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_7_2_a_engine_eval_mse.R
\name{default_params_eval_mse}
\alias{default_params_eval_mse}
\title{Default Parameters for Evaluation Engine: MSE}
\usage{
default_params_eval_mse()
}
\value{
A list of default parameters for the MSE evaluation engine.
}
\description{
Provides default parameters for the MSE evaluation engine.
}
\details{
\strong{Purpose:}
\itemize{
\item Defines engine-specific parameters that are optional but can be adjusted for specific use cases.
\item Ensures default parameters are used when none are provided in the \code{control} object.
}

\strong{Additional Parameters:}
\itemize{
\item None for this engine; it relies entirely on the base fields from the controller.
}
}
\seealso{
\code{\link[=wrapper_eval_mse]{wrapper_eval_mse()}}
}
\keyword{internal}
