% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/2_7_3_a__engine_eval_statisticalparity__mwiller.R
\name{default_params_eval_statisticalparity}
\alias{default_params_eval_statisticalparity}
\title{Default Parameters for Evaluation Engine: Statitical parity}
\usage{
default_params_eval_statisticalparity()
}
\value{
A list of default parameters for the MSE evaluation engine.
}
\description{
Provides default parameters for the MSE evaluation engine.
}
\details{
\strong{Purpose:}
\itemize{
\item Defines engine-specific parameters that are optional but can be adjusted for specific use cases.
\item Ensures default parameters are used when none are provided in the \code{control} object.
}

\strong{Additional Parameters:}
\itemize{
\item None for this engine; it relies entirely on the base fields from the controller.
}
}
\seealso{
\code{\link[=wrapper_eval_statisticalparity]{wrapper_eval_statisticalparity()}}
}
\keyword{internal}
