% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/2_5_a_engine_inprocessing_fairness_adversialdebiasing.R
\name{default_params_inprocessing_fairness_adversialdebiasing}
\alias{default_params_inprocessing_fairness_adversialdebiasing}
\title{Default Parameters for In-Processing Engines: Adversarial Debiasing}
\usage{
default_params_inprocessing_fairness_adversialdebiasing()
}
\value{
A list of default parameters for adversarial debiasing engines.
}
\description{
Provides default parameters for adversarial debiasing in-processing engines.
}
\details{
\strong{Purpose:}
\itemize{
\item Defines engine-specific parameters that are optional but can be adjusted for specific use cases.
\item These parameters are \strong{not covered by the base fields in the \code{controller_inprocessing} function}, which include:
\itemize{
\item \code{data}: The training dataset.
\item \code{protected_attributes}: Names of the protected attributes.
\item \code{target_var}: The target variable.
}
\item \strong{Additional Parameters:}
\itemize{
\item \code{main_model_params}: Parameters for the main model.
\item \code{adversary_params}: Parameters for the adversarial model.
\item \code{learning_rate}: Learning rate for adversarial optimization (default: 0.01).
\item \code{num_epochs}: Number of training epochs (default: 10).
}
\item Ensures default parameters are used when none are provided in the \code{control} object.
}
}
\seealso{
\code{\link[=wrapper_inprocessing_fairness_adversialdebiasing]{wrapper_inprocessing_fairness_adversialdebiasing()}}
}
\keyword{internal}
