% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/2_1_b__engine_split_random_stratified__mwiller.R
\name{default_params_split_random_stratified}
\alias{default_params_split_random_stratified}
\title{Default Parameters for Splitter Engine: Stratified Random Split}
\usage{
default_params_split_random_stratified()
}
\value{
A list of default parameters for the stratified random split engine.
}
\description{
Provides default parameters for the \code{split_random_stratified} engine.
}
\details{
\strong{Purpose:}
\itemize{
\item Defines optional parameters for stratified random splits.
}

\strong{Default Parameters:}
\itemize{
\item \code{split_ratio}: Proportion of data to be used for training (default: 0.7).
}
}
\seealso{
\code{\link[=wrapper_split_random_stratified]{wrapper_split_random_stratified()}}
}
\keyword{internal}
