% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_d_engine_split_userdefined.R
\name{default_params_split_userdefined}
\alias{default_params_split_userdefined}
\title{Default Parameters for Splitter Engine: User-Defined}
\usage{
default_params_split_userdefined()
}
\value{
\code{NULL}, as no parameters are needed.
}
\description{
This engine does not require or support any additional parameters because
it operates exclusively on externally provided data (via \code{control$data$train} and \code{control$data$test}).
It is intended for workflows where splitting happens outside the framework.
}
\details{
\strong{Purpose:}
\itemize{
\item Registers the engine in the global system.
\item Ensures compatibility with engine registration and validation logic.
\item Not used during execution.
}
}
\seealso{
\code{\link[=wrapper_split_userdefined]{wrapper_split_userdefined()}}
}
\keyword{internal}
