% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_4_b_engine_train_glm.R
\name{default_params_train_glm}
\alias{default_params_train_glm}
\title{Default Parameters for Training Engines: GLM}
\usage{
default_params_train_glm()
}
\value{
A list of default hyperparameters for the training engine.
}
\description{
Provides default parameters for training engines. These parameters are specific to each engine and define optional values required for execution.
}
\details{
\strong{Purpose:}
\itemize{
\item Defines engine-specific parameters that are optional but can be adjusted for specific use cases.
\item These parameters are \strong{not covered by the base fields in the \code{controller_training} function}, which include:
\itemize{
\item \code{formula}: A formula specifying the model structure.
\item \code{data}: A data frame containing the training data.
}
\item \strong{Additional Parameters:}
\itemize{
\item \code{sample_weight}: A vector of weights for the training data. Must have same length as train dataset (default: equal weights -> in Wrapper).
\item \code{family}: The family of the GLM (default: Gaussian).
}
\item Ensures default parameters are used when none are provided in the \code{control} object.
}
}
\seealso{
\code{\link[=wrapper_train_glm]{wrapper_train_glm()}}
}
\keyword{internal}
