% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_g__start_function__mwiller.R
\name{flowengineR_start}
\alias{flowengineR_start}
\title{Launch the interactive start menu for flowengineR}
\usage{
flowengineR_start()
}
\description{
This function offers an interactive menu to explore flowengineR.
}
